Remove useless cast
POWERS items are already uint64_t.

+++ cotp.c
@@ -470,7 +470,7 @@ COTPRESULT otp_generate(OTPData* data, uint64_t input, char* out_str)
 		| ((hmac[offset+3] & 0xFF)));
 	
 	static const uint64_t POWERS[] = { 1, 10, 100, 1000, 10000, 100000, 1000000, 10000000, 100000000, 1000000000 };
-	code %= (uint64_t) POWERS[data->digits];
+	code %= POWERS[data->digits];
 	
 	sprintf(out_str, "%0*" PRIu64, data->digits, code);
